requirements.txt
Django>=5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.28
django-redis>=5.4
psycopg2-binary>=2.9
python-decouple>=3.8